pyproject.toml
[project]
name = "ml-globalworth-hackathon-2025"
version = "0.1.0"
description = "Globalworth hackathon 2025"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.5.2",
    "diffusers>=0.32.2",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.2",
    "setuptools>=78.1.0",
    "torch>=2.6.0",
    "transformers>=4.50.3",
    "uvicorn>=0.34.0",
    "xformers>=0.0.29.post3",
]


[tool.setuptools.packages.find]
where = ["src"]
